refactor: POST 메서드 독립
