+++ client/src/services/apiFinancial.ts
@@ -1,4 +1,6 @@
 import { instance } from './tokenInstance';
+import { FaRecArticleReqType } from '../types/article';
+
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
 
 export const APIfinancialRecord = {
@@ -59,5 +61,33 @@ export const APIfinancialRecord = {
     const res = await instance.delete(
       `${BASE_URL}/financialrecord/${financialRecordId}/article/${financialRecordArticleId}`
     );
+  },
+  // 가계부 게시글 POST
+  postFeedArticle: async (
+    formData: FormData,
+    faRecId: number,
+    category: string,
+    faDate: Date,
+    title: string,
+    price: number,
+    content: string,
+    scope: number
+  ) => {
+    const body: FaRecArticleReqType = {
+      financialRecordId: faRecId,
+      category,
+      faDate: faDate.toISOString(),
+      title,
+      price,
+      content,
+      scope: scope === 0 ? '가계부 게시글' : '가계부 타임라인',
+    };
+    formData.append('data', JSON.stringify(body));
+
+    await instance.post(`${BASE_URL}/financialrecord/${faRecId}/article'`, formData, {
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      },
+    });
   },
 };
