+++ client/src/services/apiSns.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import { instance } from './tokenInstance';
 import { VoteType } from '../types/article';
+import { FeedArticleReqType } from '../types/article';
 
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
 
@@ -17,17 +18,31 @@ export const APISns = {
   deleteFeedArticle: async (feedArticleId: number) => {
     const res = await instance.delete(`${BASE_URL}/feedArticles/${feedArticleId}`);
   },
+  // [POST] SNS
+  postFeedArticle: async (formData: FormData, feedType: number, content: string) => {
+    const body: FeedArticleReqType = {
+      feedType: feedType === 1 ? '절약팁' : '허락해줘', // 2: 허락해줘
+      content,
+    };
+    formData.append('data', JSON.stringify(body));
+
+    await axios.post(`${BASE_URL}/feedArticles`, formData, {
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      },
+    });
+  },
 };
 
 export const APIVote = {
   // [GET] Vote
   getVote: async (feedArticleId: number) => {
-    const res = await axios.get(`${BASE_URL}/vote/${feedArticleId}`);
+    const res = await instance.get(`${BASE_URL}/vote/${feedArticleId}`);
     const voteData: VoteType = res.data;
     return voteData;
   },
   // [POST] 절약/Flex 버튼 눌렀을 시
   postVote: async (feedArticleId: number, voteType: 'SAVING' | 'FLEX') => {
-    const res = await axios.get(`${BASE_URL}/vote/${feedArticleId}?voteType=${voteType}`);
+    const res = await instance.get(`${BASE_URL}/vote/${feedArticleId}?voteType=${voteType}`);
   },
 };
